+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Add
 - Phone number indicator
 - Autonomous System Number indicator
+- Phone number and ASN indicators detection and collection
 - CVE Details resource for CVEs
 - Exploit Database resource for CVEs
 - IP Quality Score resource for phone number lookup